ReadMe.md
Physician Assistan/nurse counts the patients controlled substances
Purpose of program is to make sure the patiens pill count is correct
next step and not yet programmed is to add dates that the next 3 prescriptions can be filled on
another nicety would be different statements if the patient has more meds that supposed to
instead of negative numbers. I will do this next. 
 
To implement code, in the search bar, type ISE and click on <Run as administrator>.
This will bring up Powershell.  
At the prompt type Set-ExecutionPolicy remotesigned
Click yes to all.  
 In the View, click on script pane and Copy and paste
 code into Powershell script pane.
At the prompt in the blue command module at  c:\useres\Your Name\ 
type 
./pc.ps1 
return will run the program
in the date to be entered use mm/dd/yyyy format